requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
pandas==2.1.4
requests==2.31.0
googletrans==4.0.0-rc1
pymongo==4.6.0
python-dotenv==1.0.0
aiofiles==23.2.1
certifi
